feat: Refactor and updates to support new schema detail permalink page
