+++ src/components/elements/Icons/GitHub.tsx
@@ -1,0 +1,26 @@
+import * as React from "react";
+import { baseColors } from "../../../themes";
+
+export interface GitHubProps {
+  color?: string;
+  style?: any;
+}
+
+export const GitHub: React.FunctionComponent<GitHubProps> = (props) => {
+  const color = props.color || baseColors.black;
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      height="533.3338"
+      width="547.00061"
+      viewBox="-74.4 -120.90175 293.92166 483.60742"
+      version="1.1"
+      style={props.style}
+    >
+      <path
+        fill={color}
+        d="m -8.6516138,268.49825 c 0,2 -2.3000002,3.6 -5.2000002,3.6 -3.300002,0.3 -5.600002,-1.3 -5.600002,-3.6 0,-2 2.3,-3.6 5.200002,-3.6 3,-0.3 5.6000002,1.3 5.6000002,3.6 z m -31.1000042,-4.5 c -0.7,2 1.3,4.3 4.3,4.9 2.600002,1 5.600002,0 6.200002,-2 0.6,-2 -1.3,-4.3 -4.3,-5.2 -2.600002,-0.7 -5.500002,0.3 -6.200002,2.3 z m 44.2000076,-1.7 c -2.900012,0.7 -4.90001233,2.6 -4.60000233,4.9 0.3,2 2.90000233,3.3 5.90000233,2.6 2.9000002,-0.7 4.9000004,-2.6 4.6000004,-4.6 -0.30001,-1.9 -3.0000004,-3.2 -5.9000004,-2.9 z m 65.7999904,-383.2 c -138.7,0 -244.79998,105.3 -244.79998,244 0,110.9 69.8,205.8 169.4999866,239.2 12.800003,2.3 17.3000024,-5.6 17.3000024,-12.1 0,-6.2 -0.3,-40.4 -0.3,-61.4 0,0 -70.000009,15 -84.700011,-29.8 0,0 -11.4,-29.1 -27.799978,-36.6 0,0 -22.9,-15.7 1.599996,-15.4 0,0 24.899982,2 38.599984,25.8 21.900002,38.6 58.599997,27.5 72.899999,20.9 2.30001,-16 8.800006,-27.1 16.000001,-33.7 -55.899996,-6.2 -112.300002,-14.3 -112.300002,-110.5 0,-27.5 7.6,-41.3 23.600002,-58.9 -2.6,-6.5 -11.100002,-33.3 2.6,-67.9 20.900002,-6.5 69.00001,27 69.00001,27 19.99999,-5.6 41.49999,-8.5 62.79999,-8.5 21.3,0 42.80002,2.9 62.80002,8.5 0,0 48.1,-33.6 68.99999,-27 13.7,34.7 5.2,61.4 2.6,67.9 16.00001,17.7 25.80001,31.5 25.80001,58.9 0,96.5 -58.9,104.2 -114.8,110.5 9.2,7.9 17,22.9 17,46.4 0,33.7 -0.3,75.4 -0.3,83.6 0,6.5 4.6,14.4 17.3,12.1 99.99999,-33.2 167.79999,-128.1 167.79999,-239 0,-138.7 -112.5,-244 -251.20001,-244 z m -147.600002,344.9 c -1.3,1 -1,3.3 0.7,5.2 1.6,1.6 3.9,2.3 5.2,1 1.300002,-1 1.000002,-3.3 -0.7,-5.2 -1.6,-1.6 -3.9,-2.3 -5.2,-1 z m -10.8,-8.1 c -0.7,1.3 0.3,2.9 2.3,3.9 1.6,1 3.6,0.7 4.3,-0.7 0.7,-1.3 -0.3,-2.9 -2.3,-3.9 -2,-0.6 -3.6,-0.3 -4.3,0.7 z m 32.400002,35.6 c -1.6,1.3 -1,4.3 1.3,6.2 2.300002,2.3 5.200002,2.6 6.500002,1 1.3,-1.3 0.7,-4.3 -1.3,-6.2 -2.2,-2.3 -5.200002,-2.6 -6.500002,-1 z m -11.4,-14.7 c -1.6,1 -1.6,3.6 0,5.9 1.6,2.3 4.3,3.3 5.6,2.3 1.6,-1.3 1.6,-3.9 0,-6.2 -1.4,-2.3 -4,-3.3 -5.6,-2 z"
+      />
+    </svg>
+  );
+};
