+++ src/components/elements/Icons/Icons.stories.tsx
@@ -8,6 +8,7 @@ import {
   ContentCopy,
   CredentialCheck,
   EthrDidLogo,
+  GitHub,
   Home,
   Info,
   KeyboardArrowDown,
@@ -46,6 +47,9 @@ storiesOf("Icons", module)
   })
   .add("EthrDidLogo", () => {
     return <EthrDidLogo />;
+  })
+  .add("GitHub", () => {
+    return <GitHub />;
   })
   .add("Home", () => {
     return <Home />;
