+++ src/components/views/Schemas/index.ts
@@ -2,7 +2,9 @@ export * from "./CreateSchema";
 export * from "./SchemaCard";
 export * from "./SchemaCards";
 export * from "./SchemaDetail";
+export * from "./SchemaHeader";
 export * from "./SchemaPlayground";
+export * from "./SchemaPreview";
 export * from "./SchemaUsage";
 export * from "./SchemasTable";
 export * from "./types";
